RAG/Flask_Chatbot/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
click==8.1.6
dataclasses-json==0.5.14
faiss-cpu==1.7.4
filelock==3.12.2
Flask==2.3.2
Flask-WTF==1.1.1
frozenlist==1.4.0
fsspec==2023.6.0
greenlet==2.0.2
httptools==0.6.0
huggingface-hub==0.16.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.267
langsmith==0.0.24
MarkupSafe==2.1.3
marshmallow==3.20.1
monotonic==1.5
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numexpr==2.8.5
openai==0.27.8
openapi-schema-pydantic==1.2.4
pyarrow==11.0.0
pypdf==3.15.1
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.2
SQLAlchemy==2.0.20
tenacity==8.2.3
tiktoken==0.4.0
torch==2.0.1
transformers==4.31.0
typing-inspect==0.9.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.7
WTForms==3.0.1
yarl==1.9.2
zstandard==0.19.0
